[Kanishk] ADDED:UC7 for getting top bowling averages

+++ src/main/java/com/iplanalyserproblem/IplCSVBowler.java
@@ -34,7 +34,7 @@ public class IplCSVBowler
 	    public double average;
 
 	    @CsvBindByName(column = "Econ")
-	    public int economy;
+	    public double economy;
 
 	    @CsvBindByName(column = "SR")
 	    public double strikeRate;
